+++ cfbfastR/config.py
@@ -2,3 +2,4 @@ CFB_BASE_URL = 'https://raw.githubusercontent.com/saiemgilani/cfbfastR-data/mast
 CFB_ROSTER_URL = 'https://raw.githubusercontent.com/saiemgilani/cfbfastR-data/master/rosters/roster_{year}.parquet?raw=True'
 CFB_TEAM_LOGO_URL  = 'https://github.com/saiemgilani/cfbfastR-data/blob/master/teams_colors_logos.csv'
 CFB_TEAM_SCHEDULE_URL = 'https://raw.githubusercontent.com/saiemgilani/cfbfastR-data/master/schedules/parquet/schedules_{year}.parquet'
+CFB_TEAM_INFO_URL = 'https://raw.githubusercontent.com/saiemgilani/cfbfastR-data/master/team_info/parquet/team_info_{year}.parquet'