+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.2 Release: June 26, 2021
+- Function added: `load_cfb_pbp(years: List[int]) -> pd.DataFrame`
+- Function added: `cfb_calendar(season: int) -> pd.DataFrame`
+- Function added: `load_cfb_schedule(years: List[int]) -> pd.DataFrame`
+- Function added: `load_cfb_rosters(years: List[int]) -> pd.DataFrame`
+- Function added: `cfb_teams() -> pd.DataFrame`
+- Function added: `load_cfb_team_info(years: List[int]) -> pd.DataFrame`
+
 ## 0.0.1 Release: June 26, 2021
 - f'in round
 - findin' out